# Conflicts:
#	aergo-protobuf
#	cmd/aergosvr/aergosvr.go
#	types/rpc.pb.go

+++ tests/run_tests.sh
@@ -1,24 +1,72 @@
 # stop on errors
 set -e
+source common.sh
 
-# run the brick test
-./test-brick.sh
+arg=$1
+if [ "$arg" != "sbp" ] && [ "$arg" != "dpos" ] && [ "$arg" != "raft" ] && [ "$arg" != "brick" ]; then
+  echo "Usage: $0 [brick|sbp|dpos|raft]"
+  exit 1
+fi
+echo "Running integration tests for $arg"
 
-# delete and recreate the aergo folder
-rm -rf ./aergo-files
-mkdir aergo-files
-# copy the config file
-cp config.toml ./aergo-files/
+if [ "$arg" == "brick" ]; then
+  # run the brick test
+  ../bin/brick -V test.brick
+  exit 0
+fi
 
-# open the aergo server in testmode to create the config file
+consensus=$arg
+
+if [ "$consensus" == "sbp" ]; then
+  # delete and recreate the aergo folder
+  rm -rf ./aergo-files
+  mkdir aergo-files
+  # copy the config file
+  cp config-sbp.toml ./aergo-files/config.toml
+  # delete the old logs
+  rm -f logs
+else
+  # delete and recreate the aergo folder
+  rm -rf node1
+  rm -rf node2
+  rm -rf node3
+  mkdir node1
+  mkdir node2
+  mkdir node3
+  # copy the config files
+  cp config-node1.toml node1/config.toml
+  cp config-node2.toml node2/config.toml
+  cp config-node3.toml node3/config.toml
+  # delete the old logs
+  rm -f logs1 logs2 logs3
+  # create the genesis block
+  echo "creating genesis block..."
+  ../bin/aergosvr init --genesis ./genesis-$consensus.json --home ./node1
+  ../bin/aergosvr init --genesis ./genesis-$consensus.json --home ./node2
+  ../bin/aergosvr init --genesis ./genesis-$consensus.json --home ./node3
+fi
+
+# define the config files according to the consensus
+if [ "$consensus" == "sbp" ]; then
+  config_files=("./aergo-files/config.toml")
+elif [ "$consensus" == "dpos" ]; then
+  config_files=("./node1/config.toml" "./node2/config.toml" "./node3/config.toml")
+elif [ "$consensus" == "raft" ]; then
+  config_files=("./node1/config.toml" "./node2/config.toml" "./node3/config.toml")
+fi
+
+# define which port used for queries
+if [ "$consensus" == "sbp" ]; then
+  query_port="7845"
+else
+  query_port="9845"
+fi
+
 echo ""
-echo "starting the aergo server..."
-../bin/aergosvr --testmode --home ./aergo-files > logs 2> logs &
-pid=$!
-# wait it to be ready
-sleep 2
-
-version=$(../bin/aergocli blockchain | jq .ChainInfo.Chainid.Version | sed 's/"//g')
+echo "starting nodes..."
+start_nodes
+# wait the node(s) to be ready, expecting hardfork version 2
+wait_version 2
 
 function set_version() {
   # stop on errors
@@ -31,28 +79,18 @@ function set_version() {
   block_no=$(../bin/aergocli blockchain | jq .Height | sed 's/"//g')
   # increment 2 numbers
   block_no=$((block_no+2))
-  # terminate the server process
-  kill $pid
-  # save the hardfork config on the config file
-  echo "updating the config file..."
-  if [ $version -eq 2 ]; then
-    sed -i "s/^v2 = \"10000\"$/v2 = \"${block_no}\"/" ./aergo-files/config.toml
-  elif [ $version -eq 3 ]; then
-    sed -i "s/^v3 = \"10000\"$/v3 = \"${block_no}\"/" ./aergo-files/config.toml
-  fi
-  # restart the aergo server
-  echo "restarting the aergo server..."
-  ../bin/aergosvr --testmode --home ./aergo-files > logs 2> logs &
-  pid=$!
-  # wait it to be ready
-  sleep 3
-  # check if it worked
-  new_version=$(../bin/aergocli blockchain | jq .ChainInfo.Chainid.Version | sed 's/"//g')
-  if [ $new_version -ne $version ]; then
-    echo "Failed to change the blockchain version on the server"
-    echo "Desired: $version, Actual: $new_version"
-    exit 1
-  fi
+  # terminate the server process(es)
+  stop_nodes
+  # save the hardfork config on the config file(s)
+  echo "updating the config file(s)..."
+  for config_file in "${config_files[@]}"; do
+    sed -i "s/^v$version = \"10000\"$/v$version = \"${block_no}\"/" $config_file
+  done
+  # restart the aergo nodes
+  echo "restarting the aergo nodes..."
+  start_nodes
+  # wait the node(s) to be ready, expecting the new hardfork version
+  wait_version $version
   echo "---------------------------------"
   # do not stop on errors
   set +e
@@ -77,6 +115,10 @@ function check() {
     fi
 }
 
+# make these variables accessible to the called scripts
+export consensus
+export query_port
+
 # create the account used on tests
 echo "creating user account..."
 ../bin/aergocli account import --keystore . --if 47zh1byk8MqWkQo5y8dvbrex99ZMdgZqfydar7w2QQgQqc7YrmFsBuMeF1uHWa5TwA1ZwQ7V6 --password bmttest
@@ -87,6 +129,7 @@ check ./test-gas-op.sh
 check ./test-gas-bf.sh
 check ./test-gas-verify-proof.sh
 check ./test-gas-per-function-v2.sh
+check ./test-contract-deploy.sh
 
 # change the hardfork version
 set_version 3
@@ -98,12 +141,13 @@ check ./test-gas-op.sh
 check ./test-gas-bf.sh
 check ./test-gas-verify-proof.sh
 check ./test-gas-per-function-v3.sh
+check ./test-contract-deploy.sh
 
 # terminate the server process
 echo ""
-echo "closing the aergo server"
+echo "closing the aergo nodes"
 echo ""
-kill $pid
+stop_nodes
 
 # print the summary
 if [ $num_failed_tests -gt 0 ]; then
